Changed the way to connect
Connecting to a Docker daemon is done now by using the DockerConnectDialog. The Dock Widget with the list of daemons is being removed.
Database model also changed to mirror the way the library reads the data. Each Docker daemon you want to connect to, you should define an 'Environment'. Different versions of the API should also be catered within this app, that's why there is going to be a table to keep track of all the different versions that are being supported by the app